pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "plug-classifier"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "torch",
  "torchvision",
  "tqdm",
  "pyyaml",
  "numpy",
  "Pillow",
  "scikit-learn",
  "tensorboard",
  "rich",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
